requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
Flask==3.0.3
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
setuptools==70.3.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
Werkzeug==3.0.3
